% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drifter.R
\name{map_drifter}
\alias{map_drifter}
\title{map_drifter}
\usage{
map_drifter(x, data, xlim = c(350, 430), ylim = c(4990, 5030), raster, res = 1)
}
\arguments{
\item{x}{a fitted object of class simTidal}

\item{data}{data object created by sim_setup}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{res}{downsampling factor to product a plot faster (default = 5, full-resolution = 0)}
}
\description{
Map simulated drifter track
}
