% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{moveKcam}
\alias{moveKcam}
\title{random walk movement kernel for Campbellton River kelts}
\usage{
moveKcam(
  n = 1,
  data,
  xy = NULL,
  coa = NULL,
  dir = NULL,
  buffer = NULL,
  rho,
  a,
  b,
  shelf,
  beta
)
}
\description{
utility function not to be called by user
}
